fix(cache): Fix multiple caching issues (see details)
- Cache the fetch requests fro 4 hours (won't probably change in this
  time frame)
- Use npm instead of bun as it causes weird caching issues
