+++ frontend-nextjs/package.json
@@ -1,70 +1,70 @@
 {
-  "name": "media-impact-monitor-frontend",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "cross-env NODE_OPTIONS=--max_old_space_size=4096 next dev --turbo",
-    "build": "bun run check && next build",
-    "start": "next start",
-    "check": "bunx @biomejs/biome check --apply src"
-  },
-  "dependencies": {
-    "@custom-react-hooks/use-debounce": "^1.4.19",
-    "@custom-react-hooks/use-element-size": "^1.4.19",
-    "@custom-react-hooks/use-media-query": "1.4.19",
-    "@custom-react-hooks/use-storage": "^1.4.19",
-    "@radix-ui/react-accordion": "^1.1.2",
-    "@radix-ui/react-dialog": "1.0.5",
-    "@radix-ui/react-dropdown-menu": "2.0.6",
-    "@radix-ui/react-icons": "1.3.0",
-    "@radix-ui/react-popover": "1.0.7",
-    "@radix-ui/react-slot": "1.0.2",
-    "@radix-ui/react-tooltip": "1.0.7",
-    "@tanstack/react-query": "5.29.2",
-    "@tanstack/react-query-next-experimental": "5.29.2",
-    "@tanstack/react-table": "8.12.0",
-    "@types/bun": "1.1.1",
-    "class-variance-authority": "0.7.0",
-    "cmdk": "0.2.1",
-    "d3-scale": "4.0.2",
-    "d3-time": "3.1.0",
-    "date-fns": "3.6.0",
-    "framer-motion": "11.0.28",
-    "lucide-react": "0.341.0",
-    "next": "14.1.0",
-    "next-themes": "0.3.0",
-    "nextjs-toploader": "1.6.11",
-    "react": "18.0.0",
-    "react-day-picker": "8.10.0",
-    "react-dom": "18.0.0",
-    "recharts": "^2.12.7",
-    "seed-random": "2.2.0",
-    "slugify": "1.6.6",
-    "sonner": "1.4.41",
-    "vaul": "0.9.0",
-    "zod": "3.23.8",
-    "zustand": "^4.5.2"
-  },
-  "devDependencies": {
-    "@biomejs/biome": "1.7.2",
-    "@tailwindcss/container-queries": "0.1.1",
-    "@tailwindcss/forms": "0.5.7",
-    "@tailwindcss/typography": "0.5.12",
-    "@tanstack/react-query-devtools": "5.29.2",
-    "@types/d3-scale": "4.0.8",
-    "@types/node": "18.0.0",
-    "@types/react": "18.0.0",
-    "@types/react-dom": "18.0.0",
-    "@types/seed-random": "2.2.4",
-    "autoprefixer": "10.4.17",
-    "clsx": "2.1.0",
-    "cross-env": "7.0.3",
-    "postcss": "8.4.35",
-    "prettier": "3.2.5",
-    "sharp": "0.33.3",
-    "tailwind-merge": "2.2.1",
-    "tailwindcss": "3.4.1",
-    "tailwindcss-animate": "1.0.7",
-    "typescript": "4.7.4"
-  }
+	"name": "media-impact-monitor-frontend",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "cross-env NODE_OPTIONS=--max_old_space_size=4096 next dev --turbo",
+		"build": "bun run check && next build",
+		"start": "next start",
+		"check": "bunx @biomejs/biome check --apply src"
+	},
+	"dependencies": {
+		"@custom-react-hooks/use-debounce": "^1.4.19",
+		"@custom-react-hooks/use-element-size": "^1.4.19",
+		"@custom-react-hooks/use-media-query": "1.4.19",
+		"@custom-react-hooks/use-storage": "^1.4.19",
+		"@radix-ui/react-accordion": "^1.1.2",
+		"@radix-ui/react-dialog": "1.0.5",
+		"@radix-ui/react-dropdown-menu": "2.0.6",
+		"@radix-ui/react-icons": "1.3.0",
+		"@radix-ui/react-popover": "1.0.7",
+		"@radix-ui/react-slot": "1.0.2",
+		"@radix-ui/react-tooltip": "1.0.7",
+		"@tanstack/react-query": "5.29.2",
+		"@tanstack/react-query-next-experimental": "5.29.2",
+		"@tanstack/react-table": "8.12.0",
+		"@types/bun": "1.1.1",
+		"class-variance-authority": "0.7.0",
+		"cmdk": "0.2.1",
+		"d3-scale": "4.0.2",
+		"d3-time": "3.1.0",
+		"date-fns": "3.6.0",
+		"framer-motion": "11.0.28",
+		"lucide-react": "0.341.0",
+		"next": "14.1.0",
+		"next-themes": "0.3.0",
+		"nextjs-toploader": "1.6.11",
+		"react": "18.0.0",
+		"react-day-picker": "8.10.0",
+		"react-dom": "18.0.0",
+		"recharts": "2.12.7",
+		"seed-random": "2.2.0",
+		"slugify": "1.6.6",
+		"sonner": "1.4.41",
+		"vaul": "0.9.0",
+		"zod": "3.23.8",
+		"zustand": "^4.5.2"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "1.7.2",
+		"@tailwindcss/container-queries": "0.1.1",
+		"@tailwindcss/forms": "0.5.7",
+		"@tailwindcss/typography": "0.5.12",
+		"@tanstack/react-query-devtools": "5.29.2",
+		"@types/d3-scale": "4.0.8",
+		"@types/node": "18.0.0",
+		"@types/react": "18.0.0",
+		"@types/react-dom": "18.0.0",
+		"@types/seed-random": "2.2.4",
+		"autoprefixer": "10.4.17",
+		"clsx": "2.1.0",
+		"cross-env": "7.0.3",
+		"postcss": "8.4.35",
+		"prettier": "3.2.5",
+		"sharp": "0.33.3",
+		"tailwind-merge": "2.2.1",
+		"tailwindcss": "3.4.1",
+		"tailwindcss-animate": "1.0.7",
+		"typescript": "4.7.4"
+	}
 }
