+++ frontend-nextjs/src/components/QueryClientProvider/index.tsx
@@ -1,5 +1,6 @@
 "use client";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { ReactQueryStreamedHydration } from "@tanstack/react-query-next-experimental";
 import { useRef } from "react";
 
 function makeQueryClient() {
@@ -28,7 +29,6 @@ function getQueryClient() {
 	if (!browserQueryClient) browserQueryClient = makeQueryClient();
 	return browserQueryClient;
 }
-export const queryClient = makeQueryClient();
 
 function CustomQueryClientProvider({
 	children,
@@ -36,7 +36,7 @@ function CustomQueryClientProvider({
 	const queryClient = useRef(getQueryClient());
 	return (
 		<QueryClientProvider client={queryClient.current}>
-			{children}
+			<ReactQueryStreamedHydration>{children}</ReactQueryStreamedHydration>
 		</QueryClientProvider>
 	);
 }
